:recycle: Refactor Search into Wrapper component
