+++ frontend/lib/filterConstants.js
@@ -1,0 +1,9 @@
+export const PRICE = ['FREE', 'LOW', 'MID', 'HIGH']
+
+export const ORDER_BY = [
+  { text: 'Alphabetical', value: 0 },
+  { text: 'Highest Rating', value: 1 },
+  { text: 'Most Reviewed', value: 2 },
+]
+
+export const DIFFICULTY = ['EASY', 'MID', 'HARD', 'EXPERT']
