+++ frontend/pages/test-wrap.js
@@ -1,0 +1,20 @@
+import Wrapper from '../components/Wrapper'
+
+export default props => (
+  <Wrapper>
+    <div>
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+      <img src="https://media.giphy.com/media/PV9t4AJL5UweI/giphy.gif" />
+    </div>
+  </Wrapper>
+)
