+++ frontend/styles/components/search.scss
@@ -9,8 +9,7 @@
 }
 
 .search__bar {
-  width: 85%;
-  margin-right: 10px;
+  width: 75%;
 }
 
 .search__button {
